added a home component to be displayed at root. added some css to it. corrected the css in other components accordingly. added a confirmation to the remove button.

+++ src/app/stars/stars.component.ts
@@ -41,7 +41,9 @@ export class StarsComponent implements OnInit, OnDestroy {
   }
 
   onRemoveStar() {
-    this.starService.removeStar(this.selectedStar);
-    this.router.navigate(['stars']);
+    if (confirm('Möchten Sie wirklich diesen Eintrag löschen?')) {
+      this.starService.removeStar(this.selectedStar);
+      this.router.navigate(['stars']);
+    }
   }
 }
